Item name changes
1. ItemAPI.loadItemNames() will no longer error if materials.csv is missing from CivModCore's plugin files.
2. Custom item names no longer support lore. It will no longer be considered when setting or retrieving an item's name.
3. The ability to set custom item names has been moved onto the notifying event.

The reason for removing lore's inclusion in an item's identity is that it gets complicated very quickly. If we assume that an item's lore must match exactly, then compacted custom items will not be named correctly. If we assume that the lore must be included in order, then instances where hidden lore in injected, such as with ItemExchange, will also cause custom items to not be named correctly.